// src/types/forecast.ts
export interface ForecastItem {
  [key: string]: string | number | null | undefined;
  Tarih: string;
  Oda: number;
  Pax: number;
  Free: number;
  Yetişkin: number;
  Çocuk: number;
  "Toplam Kişi": number;
  "Yüzde%": number;
  Mevcut: number;
  "Arızalı Oda": number;
  "Arızalı Toplam": number;
  Bebek: number;
  Birleşme: number;
  "Birleşme Trace": number;
  "Blokajsız Oda": number;
  "Brut Tutar": number;
  "Comp Oda": number;
  "Day Use": number;
  "Doviz Gerçek": number;
  "Doviz Tutar": number;
  Fark_Yuzde: number;
  Forecast: number;
  "Forecast Gelir": number;
  "Forecast Kalan": number;
  "Forecast Satılan": number;
  "Forecasttan Kalan(Tek)": number;
  "Gelen Free": number;
  "Gelen Oda": number;
  "Gelen Pax": number;
  "Gelen Top Kişi": number;
  "Gelen Yetişkin": number;
  "Gelen Çocuk": number;
  "Giden Free": number;
  "Giden Oda": number;
  "Giden Pax": number;
  "Giden Toplam Kişi": number;
  "Giden Yetişkin": number;
  "Giden Çocuk": number;
  "Gun Tarih": string;
  His_For: string;
  "House Use Oda": number;
  "Info Oda": number;
  Kalan_1: number;
  Kalan_2: number;
  Kalan_3: number;
  Kalan_4: number;
  Kalan_5: number;
  Kalan_6: number;
  "Kapalı Oda": number;
  "Kdv%": number;
  "Kon.Vergisi%"?: number;
  "Kontenjan Kalan": number;
  "Kontenjan Oda": number;
  "Kontenjan Satılan": number;
  Konum1: number;
  "Konum1%": number;
  Konum2: number;
  "Konum2%": number;
  Konum3: number;
  "Konum3%": number;
  Konum4: number;
  "Konum4%": number;
  Konum5: number;
  "Konum5%": number;
  Konum6: number;
  "Mevcut(Net)": number;
  "Müşteri Tipi(1)": number;
  "Müşteri Tipi(2)": number;
  "Müşteri Tipi(3)": number;
  "Müşteri Tipi(4)": number;
  "Müşteri Tipi(5)": number;
  "Müşteri Tipi(6)": number;
  "Müşteri Tipi(7)": number;
  "Müşteri Tipi(8)": number;
  "Net Oda": number;
  NoShow: number;
  "Ort.Pax(Prm)"?: number;
  Ort_Oda_Brut: number;
  Ort_Paxp_Brut: number;
  "Otel Adı": string;
  "Otel Kodu": number;
  "Package Kdv ": number;
  "Package Kdvsiz": number;
  "Package Kon.Vergisi"?: number;
  "Package Tutar": number;
  "Pax(P)": number;
  "Pax(Prm)": number;
  "Pax(Y/C2)": number;
  "Paxp Free": number;
  "Paxp Yetişkin": number;
  "Paxp Çocuk": number;
  "Pm Sanal": number;
  "Satılan-Birleşme": number;
  "Son Durum": number;
  Son_Yuzdem: number;
  Stop: number;
  "TL Gerçek": number;
  TarihAy: string;
  "Tentative Oda": number;
  "Yatak%": number;
  "Yatak(%)": number;
  "Yatak(Mevcut)": number;
  "Yuzde%(Net)": number;
  "Yuzde%(Sondurum)": number;
}

export interface ForecastResponse {
  isSucceded: boolean;
  message: string | null;
  messageList: string[];
  value: ForecastItem[];
}

export interface ForecastDataProps {
  requestData: ForecastRequestData;
}

export interface ForecastRequestData {
  db_Id: number;
  xRez_Sirket: number;
  xBas_Tar: string;
  xBit_Tar: string;
  xtip: number;
  kon1: string;
  kon2: string;
  xchkFis_Fazla_otel_10: number;
  bas_Yil: number;
  bit_Yil: number;
  fisrci_Kapalioda_10: number;
  xRez_C_W: string;
  xSistem_Tarihi: string;
  xAlis_Tarihi: string;
  sistem_Bas1: string;
  sistem_Bit1: string;
  pmdahil_10: number;
  tip_1: string;
  xFis_Bela_tutar_10: number;
  trace_Dus_10: number;
  cev_01: null | string;
}